added search and lookup actions with corresponding views and model methods

+++ app/views/one.php
@@ -1,7 +1,11 @@
 <?php
-foreach ($data[0] as $index => $item) {
-    echo $index. ': '. $item. '<br>';
+if ($data[0]['id'] == null) {
+    echo "Nothing was found :(";
+} else {
+    foreach ($data[0] as $index => $item) {
+        echo $index . ': ' . $item . '<br>';
+    }
+    echo "<a href=\"/movie/destroy/id/" . $data[0]['id'] . " \". >DELETE ITEM</a>";
 }
-echo "<a href=\"/movie/destroy/id/". $data[0]['id']." \". >DELETE ITEM</a>";
 ?>
